Implement mouse wheel events

+++ mordant/src/jvmMain/kotlin/com/github/ajalt/mordant/internal/syscalls/jna/SyscallHandler.jna.windows.kt
@@ -322,8 +322,8 @@ internal object SyscallHandlerJnaWindows : SyscallHandlerWindows() {
             WinKernel32Lib.INPUT_RECORD.MOUSE_EVENT -> {
                 val mouseEvent = inputEvent.Event!!.MouseEvent!!
                 EventRecord.Mouse(
-                    dwMousePositionX = mouseEvent.dwMousePosition!!.X.toInt(),
-                    dwMousePositionY = mouseEvent.dwMousePosition!!.Y.toInt(),
+                    dwMousePositionX = mouseEvent.dwMousePosition!!.X,
+                    dwMousePositionY = mouseEvent.dwMousePosition!!.Y,
                     dwButtonState = mouseEvent.dwButtonState.toUInt(),
                     dwControlKeyState = mouseEvent.dwControlKeyState.toUInt(),
                     dwEventFlags = mouseEvent.dwEventFlags.toUInt(),
